+++ Program.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.Data.Sqlite;
 using LabManager.Database;
 using LabManager.Repositories;
+using LabManager.Models;
 
 var databaseConfig = new DatabaseConfig();
 new DatabaseSetup(databaseConfig);
@@ -24,24 +25,15 @@ if (modelName == "Computer")
 
     if (modelAction == "New") 
     {
-        var connection = new SqliteConnection("Data Source=database.db");
-        connection.Open();
-
         int id = Convert.ToInt32(args[2]);
         string ram = args[3];
         string processor = args[4];
 
-        var command = connection.CreateCommand();
-        command.CommandText = "INSERT INTO  Computers VALUES($id, $ram, $processor)";
-        command.Parameters.AddWithValue("$id", id);
-        command.Parameters.AddWithValue("$ram", ram);
-        command.Parameters.AddWithValue("$processor", processor);
-        command.ExecuteNonQuery();
-
-        connection.Close();
+        var computer = new Computer(id, ram, processor);
+        computerRepository.Save(computer);
     }
 
-    if (modelAction == "Show")
+    /*if (modelAction == "Show")
     {
        var connection = new SqliteConnection("Data Source=database.db");
        connection.Open(); 
@@ -112,5 +104,5 @@ if (modelName == "Computer")
 
        reader.Close();
        connection.Close();
-    }
+    }*/
 }
