+++ Repositories/ComputerRepository.cs
@@ -30,4 +30,19 @@ class ComputerRepository
         
         return computers;
     }
+
+    public void Save(Computer computer) 
+    {
+        var connection = new SqliteConnection(databaseConfig.ConnectionString);
+        connection.Open();
+
+        var command = connection.CreateCommand();
+        command.CommandText = "INSERT INTO  Computers VALUES($id, $ram, $processor)";
+        command.Parameters.AddWithValue("$id", computer.Id);
+        command.Parameters.AddWithValue("$ram", computer.Ram);
+        command.Parameters.AddWithValue("$processor", computer.Processor);
+        command.ExecuteNonQuery();
+
+        connection.Close();
+    }
 }
